update for neovim

+++ run_after_install-nvim-config.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
 
-ln -s ~/projects/personal/dotfiles/nvim ~/.config/nvim/lua/custom
+if [[ ! -L ~/.config/nvim/lua/custom ]]; then
+  ln -s ~/projects/personal/dotfiles/nvim ~/.config/nvim/lua/custom
+fi
